gas optimization(30K savings): make cellToFire uint256

+++ src/generals/ConsecutiveShooterGeneral.sol
@@ -21,10 +21,10 @@ contract ConsecutiveShooterGeneral is IGeneral {
         uint192, /* myBoard */
         uint192, /* myAttacks */
         uint192, /* opponentsAttacks */
-        uint8 myLastMove,
-        uint8, /* opponentsLastMove */
+        uint256 myLastMove,
+        uint256, /* opponentsLastMove */
         uint64 /* opponentsDiscoveredFleet */
-    ) external pure returns (uint8) {
+    ) external pure returns (uint256) {
         if (myLastMove == 255) {
             // game just started
             return 0;
